api/requirements_fastapi_fixed.txt
# Minimal FastAPI requirements for Python 3.13
# Using latest versions that work with Python 3.13

alphagenome 
numpy
pandas
jupyter
matplotlib

# Core FastAPI stack
fastapi>=0.110.0
uvicorn[standard]>=0.27.0

# Environment and HTTP
python-dotenv
requests

